513 support, switching from PolyHook to urmem (#3)
* Start on 513 support

* e

* All windows sigs now work on 513

* Fix that annoyance

* 100% untested

* More linux stuff

* Linux GSTI

* screaming

* QQQQQQQQ

* I have no clue what I did but it works now

* Eugh

* Wow, I can move all of Windows over to urmem AND IT WORKS PERFECTLY AAAAAAAAAAAAA

* I'm considering just saying "good luck if you want anything but FFI on Linux".

* Srsly

* regparm3

* Fix WIndows compile

* callGlobalProc is regparm3

* Bleh

* Bleh

+++ byond-extools/src/core/proc_management.h
@@ -16,7 +16,7 @@ namespace Core
 		Proc(std::string name);
 		Proc(unsigned int id);
 		std::string name;
-		short id;
+		unsigned int id;
 
 		ProcArrayEntry* proc_table_entry;
 		ProcSetupEntry* setup_entry_bytecode;
